py_files/INTERVIEW/requirements.txt
groq>=0.14.0
python-dotenv==1.0.0
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.26.0
pyttsx3==2.90
PyPDF2==3.0.1
docx2txt==0.8
